#dodat deo za transakcije

+++ IzdajMe/src/main/java/com/izdajMe/izdajMe/services/AccountDeleteRequestServiceImpl.java
@@ -8,6 +8,7 @@ import com.izdajMe.izdajMe.repository.ConcurentWatcherRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.mail.SimpleMailMessage;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -41,12 +42,23 @@ public class AccountDeleteRequestServiceImpl implements AccountDeleteRequestServ
         return allRequestsList;
     }
 
+    @Transactional(readOnly = false)
     public Boolean acceptRequest(AccountDeleteRequest accountDeleteRequest) {
-        accountDeleteRequestRepository.deleteById(accountDeleteRequest.getId());
-        sendNotificationForAcceptingRequest(accountDeleteRequest);
-        return true;
+        if(!concurentWatcherRepository.findByTableName("AnswerToDeclineRequest").getWriting()) {
+            ConcurentWatcher cw = concurentWatcherRepository.findByTableName("AnswerToDeclineRequest");
+            cw.setWriting(true);
+            concurentWatcherRepository.save(cw);
+
+            accountDeleteRequestRepository.deleteById(accountDeleteRequest.getId());
+            sendNotificationForAcceptingRequest(accountDeleteRequest);
+            cw.setWriting(false);
+            concurentWatcherRepository.save(cw);
+            return true;
+        } else
+            return false;
     }
 
+    @Transactional(readOnly = false)
     public Boolean declineRequest(AccountDeleteRequest accountDeleteRequest) {
         if(!concurentWatcherRepository.findByTableName("AnswerToDeclineRequest").getWriting()) {
             ConcurentWatcher cw = concurentWatcherRepository.findByTableName("AnswerToDeclineRequest");
